move utility to the utils
